Add Channel Working!
Refactored the API to use Channel and Video classes to make things more OOP. Procedural was getting a little out of hand lol.

Channels are now added with a default of the last 120 videos for the channel.

Still need to insert channel styles, topics, etc. Also need to give feedback to user and clear form on success.